Moved Kafka consumer to separate ElasticSearch service
