+++ elasticsearch_service/go.mod
@@ -1,0 +1,5 @@
+module internship_project/elasticsearch_service
+
+go 1.15
+
+require github.com/segmentio/kafka-go v0.4.8
