+++ elasticsearch_service/main.go
@@ -1,0 +1,10 @@
+package main
+
+import (
+	elkafka "internship_project/elasticsearch_service/elastic_kafkahelpers"
+)
+
+func main() {
+	kafkaConsumer := elkafka.NewConsumer("ava-internship")
+	kafkaConsumer.Consume()
+}
