+++ kafkahelpers/getters.go
@@ -18,24 +18,6 @@ func GetConnection() *kafka.Conn {
 	return conn
 }
 
-func NewConsumer(topicName string) KafkaConsumer {
-	r := kafka.NewReader(kafka.ReaderConfig{
-		Brokers:   []string{"localhost:9092"},
-		Topic:     topicName,
-		Partition: 0,
-		MinBytes:  10e2, // 10KB
-		MaxBytes:  10e6, // 10MB
-	})
-
-	r.SetOffset(0)
-
-	consumer := KafkaConsumer{
-		Reader: r,
-	}
-
-	return consumer
-}
-
 func GetWriter(topicName string) *kafka.Writer {
 	w := kafka.NewWriter(kafka.WriterConfig{
 		Brokers:  []string{"localhost:9092"},
